requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
connected-components-3d==3.12.4
contourpy==1.1.1
cycler==0.12.1
fonttools==4.47.0
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.7.4
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
tifffile==2023.7.10
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.9.0
urllib3==2.1.0
zipp==3.17.0
